Trying to solve the position of the corners now
(This requires keeping edges in the same place.)

Note that rotation is done using complex numbers, in theory, but I found
that the search needed to have too many levels of depth to get anywhere
useful, so I disabled that code by taking numpy.abs() of my transforms.

I managed to find one useful -ish transform using this method. It is the
rrtttrrtttt one.

I might try to:
* turn my search from depth-first to breadth-first
* add rot3 ** -1 and turn ** -1
* make depth increase, and have a separate depth_limit variable
** increase depth_limit whenever abs(transform[:4,:4]) == identity(4)
* use complex numbers for transforms, but abs when storing in found
** apart from the special case of edges == identity?

... yeah, so I need to look at some kind of priority-based search algorithm
with the limit being number of searched items rather than depth, eh?